requirement.txt
accelerate==0.27.2
bitsandbytes==0.43.0
datasets==2.18.0
huggingface-hub==0.22.1
omegaconf==2.3.0
hydra-core==1.3.2
pandas==2.2.1
peft==0.9.0
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
transformers==4.39.1
